+++ src/main/java/kr/santanrudolph/everyvent/global/config/RedisConnectionConfig.java
@@ -1,0 +1,29 @@
+package kr.santanrudolph.everyvent.global.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+
+@Configuration
+public class RedisConnectionConfig {
+
+  @Value("${spring.data.redis.host}")
+  private String host;
+
+  @Value("${spring.data.redis.port}")
+  private int port;
+
+  @Bean
+  public RedisConnectionFactory redisConnectionFactory() {
+    RedisStandaloneConfiguration config = new RedisStandaloneConfiguration();
+    config.setHostName(host);
+    config.setPort(port);
+    // TODO: 운영 환경에서는 password 설정 필요
+    // config.setPassword(RedisPassword.of(password));
+
+    return new LettuceConnectionFactory(config);
+  }
+}
